requirements.txt
streamlit
python-dotenv
pandas
pypdf
sentence-transformers
openpyxl
numpy
requests
